HC-12834 Open up CORS headers
This isn't the ideal solution -- we should be checking against a
whitelist of allowed domains, or *.hipchat.com if on Cloud/SaaS.  But
for BTF we would want to check against the public DNS name of the
server, which would require adding logic to read configuration
(currently in txhipchat).  Too much for the current issue; this is
at least no worse than what was there previously.